Move UI files into own package
