--- hypedyn-ui/src/main/scala/org/narrativeandplay/hypedyn/package.scala
@@ -1,5 +1,0 @@
-package org.narrativeandplay
-
-package object hypedyn {
-
-}
