+++ hypedyn-ui/src/main/scala/org/narrativeandplay/hypedyn/ui/utils/ExpandableEmptySpace.scala
@@ -1,0 +1,7 @@
+package org.narrativeandplay.hypedyn.ui.utils
+
+import scalafx.scene.layout.{Priority, HBox}
+
+class ExpandableEmptySpace extends HBox {
+  HBox.setHgrow(this, Priority.Always)
+}
