--- hypedyn-ui/src/main/scala/org/narrativeandplay/hypedyn/utils/ExpandableEmptySpace.scala
@@ -1,7 +1,0 @@
-package org.narrativeandplay.hypedyn.utils
-
-import scalafx.scene.layout.{Priority, HBox}
-
-class ExpandableEmptySpace extends HBox {
-  HBox.setHgrow(this, Priority.Always)
-}
